Pranav - Dataset Image Reduction
Found out how to reduce the images and save them into new folders which it the file convertImages.py. I also added a couple of other files or .py in order to clean up and make the repo neater. I assigned both collaborators there own .py files to experiement on and main.py is the final test python file for the model. BackupLarxel isn't really tht useful just a backup of failed code.